requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
Flask-Caching==2.3.0
passlib
marshmallow
python-dotenv
flask-migrate
gunicorn
redis
rq
psycopg2
requests